requirements.txt
discord.py==2.3.1
stripe==5.0.0
requests==2.31.0
